Fix issue where placing a piece on a square allowed overwriting it later
Previously, it was possible to place a piece on a square and then overwrite it later in the game. This commit fixes this issue by preventing pieces from being placed on squares where they already exist. This behavior is more in line with the expected behavior of the game, where pieces should not be allowed to be placed on already-occupied squares.